+++ src/Splitter.h
@@ -7,10 +7,14 @@ struct Splitter {
   Key from;
   Key to;
   Key tp_key;
-  int n_particles;
+  // 64-bit: this is a PARTICLE COUNT, and on the first pass of
+  // OctDecomposition::findSplitters the candidate range is the whole
+  // snapshot. int truncated 24461180928 to -1308622848 and the range was
+  // then never subdivided (job 5332118, 2026-08-23).
+  long n_particles;
 
   Splitter() {}
-  Splitter(Key from_, Key to_, Key tp_key_, int n_particles_) :
+  Splitter(Key from_, Key to_, Key tp_key_, long n_particles_) :
     from(from_), to(to_), tp_key(tp_key_), n_particles(n_particles_) {}
 
   void pup(PUP::er &p) {
